+++ src/validation.cpp
@@ -48,6 +48,7 @@
 #include "util.h"
 #include "utilmoneystr.h"
 #include "validationinterface.h"
+#include "warnings.h"
 #include "zpivchain.h"
 #include "zpiv/zerocoin.h"
 #include "zpiv/zpivmodule.h"
